requirements.txt
fastapi
uvicorn
pydantic
huggingface-hub
python-dotenv
requests
PyPDF2
